requirements.txt
Django==1.11.1
djangorestframework==3.6.3
factory-boy==2.8.1
flake8
psycopg2
